# src/main/java/com/notifications/dto/request/RegisterRequest.java
package com.notifications.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.Data;

@Data
public class RegisterRequest {
    @NotBlank @Email @Size(max = 320)
    private String email;

    @NotBlank @Size(min = 8, max = 72)
    private String password;
}